removed dummy nav items, integrated ops list component everywhere
